refactor(engine): decompose buffer_controller.rs into submodules (SPL-14 #2436)
