--- crates/engine/src/local_copy/buffer_pool/buffer_controller.rs
@@ -1,2129 +1,0 @@
-//! PID-style buffer-size controller.
-//!
-//! Implements the proportional-integral-derivative controller described in
-//! `docs/design/adaptive-buffer-controller.md`. The controller lives in the
-//! buffer pool subsystem alongside [`super::throughput`] which provides the
-//! EMA-based throughput signal the controller consumes.
-//!
-//! The controller is wired into [`super::BufferPool`] via the
-//! [`BufferPool::with_buffer_controller`](super::BufferPool::with_buffer_controller)
-//! builder method. When enabled, [`BufferPool::record_transfer`](super::BufferPool::record_transfer)
-//! feeds throughput samples to the controller, and
-//! [`BufferPool::recommended_buffer_size`](super::BufferPool::recommended_buffer_size)
-//! returns the controller's PID-driven recommendation instead of the
-//! simpler EMA-based heuristic.
-//!
-//! Ziegler-Nichols closed-loop tuning is documented in section 6 of the
-//! design doc; default gains shipped here come from that table.
-//!
-//! # Loop body
-//!
-//! ```text
-//! e        = setpoint - throughput
-//! P        = K_p * e
-//! I       += K_i * e * dt    (clamped to anti-windup window)
-//! D        = K_d * (e - e_prev) / dt
-//! u_next   = clamp(u_prev + (P + I + D) * scale, min_size, max_size)
-//! ```
-//!
-//! `scale` is `1.0` because the unit analysis works out: `e` is in bytes per
-//! second, `dt` is in seconds, so `K_i * e * dt` is in bytes; `K_p * e` and
-//! `K_d * (e - e_prev) / dt` are also in bytes once the dimensionless gains
-//! are applied. The accumulator therefore moves the buffer size by an amount
-//! whose units match the manipulated variable. See section 3.2 of the RFC.
-
-use std::sync::Mutex;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::time::{Duration, Instant};
-
-/// Default proportional gain (Ziegler-Nichols LAN preset).
-const DEFAULT_GAIN_P: f64 = 0.6;
-/// Default integral gain (Ziegler-Nichols LAN preset).
-const DEFAULT_GAIN_I: f64 = 0.2;
-/// Default derivative gain (Ziegler-Nichols LAN preset).
-const DEFAULT_GAIN_D: f64 = 0.05;
-
-/// Default minimum buffer size (16 KiB), per RFC section 3.1.
-const DEFAULT_MIN_SIZE: usize = 16 * 1024;
-/// Default maximum buffer size (4 MiB), per RFC section 3.1.
-const DEFAULT_MAX_SIZE: usize = 4 * 1024 * 1024;
-
-/// Default sample interval used to seed the controller before the first
-/// `observe` call (100 ms, matching the RFC section 3.3 time trigger).
-const DEFAULT_SAMPLE_INTERVAL_MS: u64 = 100;
-
-/// Lower bound on `dt` (1 ms) to avoid divide-by-zero and amplifying noise
-/// from sub-millisecond clock jitter on the derivative term.
-const MIN_DT: Duration = Duration::from_millis(1);
-/// Upper bound on `dt` (5 s) to avoid integral windup when the producer has
-/// stalled for an extended period.
-const MAX_DT: Duration = Duration::from_secs(5);
-
-/// Builder-style configuration for [`AdaptiveBufferController`].
-///
-/// All fields are advisory; defaults match the LAN preset from
-/// `docs/design/adaptive-buffer-controller.md` section 6.
-#[derive(Debug, Clone, Copy)]
-pub struct ControllerConfig {
-    setpoint_bytes_per_sec: u64,
-    gain_p: f64,
-    gain_i: f64,
-    gain_d: f64,
-    anti_windup_clamp: (f64, f64),
-    min_size: usize,
-    max_size: usize,
-    sample_interval_ms: u64,
-}
-
-impl ControllerConfig {
-    /// Constructs a config seeded with Ziegler-Nichols LAN-preset gains.
-    ///
-    /// `setpoint_bytes_per_sec` is the target throughput in bytes per second.
-    /// The default anti-windup window is `+/- max_size` bytes, matching the
-    /// RFC section 3.4 guidance that `K_i * I_max` cannot exceed half the
-    /// buffer-size range.
-    #[must_use]
-    pub fn new(setpoint_bytes_per_sec: u64) -> Self {
-        let max_size = DEFAULT_MAX_SIZE;
-        Self {
-            setpoint_bytes_per_sec,
-            gain_p: DEFAULT_GAIN_P,
-            gain_i: DEFAULT_GAIN_I,
-            gain_d: DEFAULT_GAIN_D,
-            anti_windup_clamp: (-(max_size as f64), max_size as f64),
-            min_size: DEFAULT_MIN_SIZE,
-            max_size,
-            sample_interval_ms: DEFAULT_SAMPLE_INTERVAL_MS,
-        }
-    }
-
-    /// Overrides the lower clamp on buffer size.
-    #[must_use]
-    pub fn min_size(mut self, n: usize) -> Self {
-        self.min_size = n.max(1);
-        self
-    }
-
-    /// Overrides the upper clamp on buffer size.
-    #[must_use]
-    pub fn max_size(mut self, n: usize) -> Self {
-        self.max_size = n.max(1);
-        if self.anti_windup_clamp == (-(DEFAULT_MAX_SIZE as f64), DEFAULT_MAX_SIZE as f64) {
-            self.anti_windup_clamp = (-(self.max_size as f64), self.max_size as f64);
-        }
-        self
-    }
-
-    /// Overrides the proportional, integral, and derivative gains.
-    #[must_use]
-    pub fn gains(mut self, p: f64, i: f64, d: f64) -> Self {
-        self.gain_p = p;
-        self.gain_i = i;
-        self.gain_d = d;
-        self
-    }
-
-    /// Overrides the anti-windup integral clamp window as `(low, high)`.
-    #[must_use]
-    pub fn anti_windup_clamp(mut self, lo_hi: (f64, f64)) -> Self {
-        self.anti_windup_clamp = lo_hi;
-        self
-    }
-
-    /// Overrides the default sample interval used to seed the controller.
-    #[must_use]
-    pub fn sample_interval_ms(mut self, ms: u64) -> Self {
-        self.sample_interval_ms = ms.max(1);
-        self
-    }
-
-    /// Builds an [`AdaptiveBufferController`] from this configuration.
-    ///
-    /// The initial buffer size is set to the midpoint of `[min_size, max_size]`
-    /// rounded down, so the controller has equal room to grow or shrink in
-    /// response to the first sample.
-    #[must_use]
-    pub fn build(self) -> AdaptiveBufferController {
-        debug_assert!(self.min_size <= self.max_size);
-        let initial = self.min_size.saturating_add(
-            self.max_size
-                .saturating_sub(self.min_size)
-                .saturating_div(2),
-        );
-        AdaptiveBufferController {
-            config: self,
-            buffer_size: AtomicUsize::new(initial.clamp(self.min_size, self.max_size)),
-            state: Mutex::new(ControllerState::default()),
-        }
-    }
-}
-
-/// Inner mutable state guarded by a single mutex.
-///
-/// Contains the integrator, the previous error sample, and the timestamp of
-/// the previous sample. A single mutex is used because all three values are
-/// updated atomically on every call to [`AdaptiveBufferController::observe`].
-#[derive(Debug, Default)]
-struct ControllerState {
-    integral: f64,
-    prev_error: f64,
-    prev_sample_at: Option<Instant>,
-}
-
-/// PID-style buffer-size controller.
-///
-/// The controller is `Send + Sync`. The current recommended buffer size is
-/// stored in an [`AtomicUsize`] for lock-free reads from the hot path; the
-/// PID state is updated under a [`Mutex`] only on the (relatively rare)
-/// sample interval.
-#[derive(Debug)]
-pub struct AdaptiveBufferController {
-    config: ControllerConfig,
-    buffer_size: AtomicUsize,
-    state: Mutex<ControllerState>,
-}
-
-impl AdaptiveBufferController {
-    /// Returns the current recommended buffer size.
-    #[must_use]
-    pub fn buffer_size(&self) -> usize {
-        self.buffer_size.load(Ordering::Relaxed)
-    }
-
-    /// Returns the configured setpoint (bytes per second).
-    #[must_use]
-    pub fn setpoint(&self) -> u64 {
-        self.config.setpoint_bytes_per_sec
-    }
-
-    /// Feeds an observed throughput sample (bytes per second since the last
-    /// sample) and returns the new recommended buffer size.
-    pub fn observe(&self, throughput_bps: u64) -> usize {
-        self.observe_at(throughput_bps, Instant::now())
-    }
-
-    /// Deterministic testing hook: feeds a sample with an explicit timestamp.
-    pub(crate) fn observe_at(&self, throughput_bps: u64, now: Instant) -> usize {
-        let mut state = match self.state.lock() {
-            Ok(guard) => guard,
-            Err(poisoned) => poisoned.into_inner(),
-        };
-
-        let dt = compute_dt(state.prev_sample_at, now, self.config.sample_interval_ms);
-        let dt_secs = dt.as_secs_f64();
-
-        let setpoint = self.config.setpoint_bytes_per_sec as f64;
-        let measured = throughput_bps as f64;
-        let error = setpoint - measured;
-
-        let p_term = self.config.gain_p * error;
-
-        let raw_integral = state.integral + self.config.gain_i * error * dt_secs;
-        let (lo, hi) = self.config.anti_windup_clamp;
-        let clamped_integral = clamp_f64(raw_integral, lo, hi);
-        state.integral = clamped_integral;
-
-        let derivative = if state.prev_sample_at.is_some() {
-            self.config.gain_d * (error - state.prev_error) / dt_secs
-        } else {
-            // First sample: no derivative history yet.
-            0.0
-        };
-
-        let output = p_term + clamped_integral + derivative;
-
-        let current = self.buffer_size.load(Ordering::Relaxed) as f64;
-        let scale = 1.0;
-        let next_raw = current + output * scale;
-        let next = clamp_f64(
-            next_raw,
-            self.config.min_size as f64,
-            self.config.max_size as f64,
-        );
-        let next_size = next as usize;
-        let next_size = next_size.clamp(self.config.min_size, self.config.max_size);
-
-        self.buffer_size.store(next_size, Ordering::Relaxed);
-
-        state.prev_error = error;
-        state.prev_sample_at = Some(now);
-
-        next_size
-    }
-
-    /// Resets the integrator and derivative history.
-    ///
-    /// Called on protocol renegotiation, pipeline restart, or any other event
-    /// described in section 3.5 of the RFC. The recommended buffer size is
-    /// preserved; only the PID accumulators are zeroed.
-    pub fn reset(&self) {
-        let mut state = match self.state.lock() {
-            Ok(guard) => guard,
-            Err(poisoned) => poisoned.into_inner(),
-        };
-        state.integral = 0.0;
-        state.prev_error = 0.0;
-        state.prev_sample_at = None;
-    }
-
-    /// Returns the configured minimum buffer size.
-    #[must_use]
-    pub fn min_size(&self) -> usize {
-        self.config.min_size
-    }
-
-    /// Returns the configured maximum buffer size.
-    #[must_use]
-    pub fn max_size(&self) -> usize {
-        self.config.max_size
-    }
-}
-
-/// Computes a bounded `dt` from the previous and current sample timestamps.
-///
-/// On the very first sample (no previous timestamp) this returns the
-/// configured `sample_interval_ms`. Otherwise the elapsed duration is
-/// clamped to `[MIN_DT, MAX_DT]` to avoid divide-by-zero on the derivative
-/// term and integrator windup on a stalled producer.
-fn compute_dt(prev: Option<Instant>, now: Instant, default_ms: u64) -> Duration {
-    let raw = match prev {
-        Some(t) => now.saturating_duration_since(t),
-        None => Duration::from_millis(default_ms.max(1)),
-    };
-    if raw < MIN_DT {
-        MIN_DT
-    } else if raw > MAX_DT {
-        MAX_DT
-    } else {
-        raw
-    }
-}
-
-/// Clamps an `f64` to `[lo, hi]`, treating NaN as `lo`.
-///
-/// Using a free function avoids relying on `f64::clamp` semantics for NaN,
-/// which panic in debug builds when `lo > hi`.
-fn clamp_f64(v: f64, lo: f64, hi: f64) -> f64 {
-    if v.is_nan() {
-        return lo;
-    }
-    if v < lo {
-        lo
-    } else if v > hi {
-        hi
-    } else {
-        v
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn lan_setpoint() -> u64 {
-        // 100 MB/s, a representative LAN target.
-        100 * 1024 * 1024
-    }
-
-    fn fresh(setpoint: u64) -> AdaptiveBufferController {
-        ControllerConfig::new(setpoint).build()
-    }
-
-    fn fresh_small_window() -> AdaptiveBufferController {
-        ControllerConfig::new(lan_setpoint())
-            .min_size(16 * 1024)
-            .max_size(1024 * 1024)
-            .build()
-    }
-
-    #[test]
-    fn proportional_term_grows_buffer_below_setpoint() {
-        let ctrl = fresh_small_window();
-        let before = ctrl.buffer_size();
-        let now = Instant::now();
-        // Throughput well under setpoint -> error positive -> buffer grows.
-        let after = ctrl.observe_at(1024, now + Duration::from_millis(100));
-        assert!(
-            after > before,
-            "buffer should grow when below setpoint: before={before}, after={after}"
-        );
-    }
-
-    #[test]
-    fn proportional_term_shrinks_buffer_above_setpoint() {
-        let ctrl = fresh_small_window();
-        let before = ctrl.buffer_size();
-        let now = Instant::now();
-        // Throughput well above setpoint -> error negative -> buffer shrinks.
-        let after = ctrl.observe_at(lan_setpoint() * 4, now + Duration::from_millis(100));
-        assert!(
-            after < before,
-            "buffer should shrink when above setpoint: before={before}, after={after}"
-        );
-    }
-
-    #[test]
-    fn integral_term_eliminates_steady_state_error_after_repeated_low_samples() {
-        // Pure-P would leave a residual offset; the integrator should keep
-        // pushing the buffer up across repeated identical low samples.
-        // Use a moderate setpoint (1 MB/s) so the integral accumulates
-        // gradually without saturating the anti-windup clamp or hitting
-        // max_size on the first step.
-        let ctrl = ControllerConfig::new(1_000_000)
-            .gains(0.0, 0.5, 0.0)
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-        let mut now = Instant::now();
-        let mut last = ctrl.buffer_size();
-        let mut grew_each_step = true;
-        for _ in 0..5 {
-            now += Duration::from_millis(100);
-            let next = ctrl.observe_at(0, now);
-            if next <= last {
-                grew_each_step = false;
-                break;
-            }
-            last = next;
-        }
-        assert!(
-            grew_each_step,
-            "integral term should accumulate and keep growing the buffer"
-        );
-    }
-
-    #[test]
-    fn derivative_term_dampens_overshoot_on_step_input() {
-        let setpoint = lan_setpoint();
-        let no_d = ControllerConfig::new(setpoint)
-            .gains(0.6, 0.2, 0.0)
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-        let with_d = ControllerConfig::new(setpoint)
-            .gains(0.6, 0.2, 0.05)
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-        let t0 = Instant::now();
-        // Step from far-below-setpoint (large positive error) to far-above
-        // (large negative error). The derivative term should temper the
-        // resulting swing relative to a P+I-only controller.
-        no_d.observe_at(0, t0 + Duration::from_millis(100));
-        with_d.observe_at(0, t0 + Duration::from_millis(100));
-        let no_d_after = no_d.observe_at(setpoint * 4, t0 + Duration::from_millis(200));
-        let with_d_after = with_d.observe_at(setpoint * 4, t0 + Duration::from_millis(200));
-        // Both swing downward on the second sample (error flipped sign and
-        // grew rapidly); the derivative-equipped controller swings less far.
-        let no_d_delta = no_d_after as i64 - no_d.config.max_size as i64 / 2;
-        let with_d_delta = with_d_after as i64 - with_d.config.max_size as i64 / 2;
-        assert!(
-            with_d_delta.abs() < no_d_delta.abs() || with_d_after >= no_d_after,
-            "derivative should damp overshoot: no_d_after={no_d_after}, with_d_after={with_d_after}"
-        );
-    }
-
-    #[test]
-    fn anti_windup_clamps_integral_under_sustained_error() {
-        let cap = 1_000.0;
-        let ctrl = ControllerConfig::new(lan_setpoint())
-            .gains(0.0, 1.0, 0.0)
-            .anti_windup_clamp((-cap, cap))
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-        let mut now = Instant::now();
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(0, now);
-        }
-        let state = ctrl.state.lock().unwrap();
-        assert!(
-            state.integral <= cap + f64::EPSILON,
-            "integral should be clamped to the anti-windup ceiling, got {}",
-            state.integral
-        );
-        assert!(
-            state.integral >= -cap - f64::EPSILON,
-            "integral should be clamped to the anti-windup floor, got {}",
-            state.integral
-        );
-    }
-
-    #[test]
-    fn clamp_to_min_size_on_persistent_overshoot() {
-        let min = 16 * 1024;
-        let ctrl = ControllerConfig::new(lan_setpoint())
-            .min_size(min)
-            .max_size(1024 * 1024)
-            .build();
-        let mut now = Instant::now();
-        // Massive sustained over-target -> error very negative -> buffer
-        // should bottom out at min_size.
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(lan_setpoint() * 100, now);
-        }
-        assert_eq!(ctrl.buffer_size(), min);
-    }
-
-    #[test]
-    fn clamp_to_max_size_on_persistent_undershoot() {
-        let max = 1024 * 1024;
-        let ctrl = ControllerConfig::new(lan_setpoint())
-            .min_size(16 * 1024)
-            .max_size(max)
-            .build();
-        let mut now = Instant::now();
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(0, now);
-        }
-        assert_eq!(ctrl.buffer_size(), max);
-    }
-
-    #[test]
-    fn reset_clears_integrator_and_derivative_history() {
-        let ctrl = fresh(lan_setpoint());
-        let mut now = Instant::now();
-        for _ in 0..10 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(0, now);
-        }
-        {
-            let state = ctrl.state.lock().unwrap();
-            assert!(state.integral > 0.0);
-            assert!(state.prev_sample_at.is_some());
-        }
-        ctrl.reset();
-        let state = ctrl.state.lock().unwrap();
-        assert_eq!(state.integral, 0.0);
-        assert_eq!(state.prev_error, 0.0);
-        assert!(state.prev_sample_at.is_none());
-    }
-
-    #[test]
-    fn zero_dt_does_not_panic_or_divide_by_zero() {
-        let ctrl = fresh(lan_setpoint());
-        let now = Instant::now();
-        // Two samples at the same instant -> raw dt = 0; clamp must apply.
-        let first = ctrl.observe_at(1024, now);
-        let second = ctrl.observe_at(2048, now);
-        assert!(first >= ctrl.config.min_size);
-        assert!(second >= ctrl.config.min_size);
-        assert!(first <= ctrl.config.max_size);
-        assert!(second <= ctrl.config.max_size);
-    }
-
-    #[test]
-    fn ziegler_nichols_default_gains_converge_within_50_samples_to_within_10_percent() {
-        // Synthetic plant: throughput is proportional to the buffer size,
-        // saturating at the link capacity. The buffer-size unit is bytes,
-        // so we use a proportionality constant chosen so that the link
-        // capacity is reached when the buffer is mid-range.
-        let setpoint = lan_setpoint();
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-        // Pick k so that buffer_size = max_size / 2 yields throughput =
-        // setpoint exactly. Then the controller's optimum sits at the
-        // mid-range default initial size, which is within scope for the
-        // proportional-only first step.
-        let mid = (16 * 1024 + 4 * 1024 * 1024) / 2;
-        let k = setpoint as f64 / mid as f64;
-        let cap = setpoint as f64 * 4.0;
-        let mut now = Instant::now();
-        let mut converged = false;
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            let buf = ctrl.buffer_size() as f64;
-            let throughput = (k * buf).min(cap);
-            ctrl.observe_at(throughput as u64, now);
-            let buf_now = ctrl.buffer_size() as f64;
-            let measured = (k * buf_now).min(cap);
-            if (measured - setpoint as f64).abs() / setpoint as f64 <= 0.10 {
-                converged = true;
-                break;
-            }
-        }
-        assert!(
-            converged,
-            "controller failed to converge within 50 samples (final buffer={} bytes)",
-            ctrl.buffer_size()
-        );
-    }
-
-    #[test]
-    fn config_builder_defaults_match_rfc() {
-        let cfg = ControllerConfig::new(1_000);
-        assert_eq!(cfg.gain_p, DEFAULT_GAIN_P);
-        assert_eq!(cfg.gain_i, DEFAULT_GAIN_I);
-        assert_eq!(cfg.gain_d, DEFAULT_GAIN_D);
-        assert_eq!(cfg.min_size, DEFAULT_MIN_SIZE);
-        assert_eq!(cfg.max_size, DEFAULT_MAX_SIZE);
-    }
-
-    #[test]
-    fn observe_returns_size_within_bounds() {
-        let ctrl = fresh_small_window();
-        let mut now = Instant::now();
-        for sample in [0u64, 1, lan_setpoint(), lan_setpoint() * 10, u64::MAX / 2] {
-            now += Duration::from_millis(50);
-            let size = ctrl.observe_at(sample, now);
-            assert!(size >= ctrl.config.min_size);
-            assert!(size <= ctrl.config.max_size);
-        }
-    }
-
-    #[test]
-    fn reset_preserves_buffer_size() {
-        let ctrl = fresh(lan_setpoint());
-        let mut now = Instant::now();
-        for _ in 0..5 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(0, now);
-        }
-        let before = ctrl.buffer_size();
-        ctrl.reset();
-        assert_eq!(
-            ctrl.buffer_size(),
-            before,
-            "reset should preserve buffer size, only clearing PID accumulators"
-        );
-    }
-
-    #[test]
-    fn min_size_accessor() {
-        let ctrl = ControllerConfig::new(1_000).min_size(8 * 1024).build();
-        assert_eq!(ctrl.min_size(), 8 * 1024);
-    }
-
-    #[test]
-    fn max_size_accessor() {
-        let ctrl = ControllerConfig::new(1_000).max_size(512 * 1024).build();
-        assert_eq!(ctrl.max_size(), 512 * 1024);
-    }
-
-    // --- Convergence tests (task #2096) ---
-    //
-    // These tests verify that the PID controller converges to optimal
-    // buffer sizes under varying synthetic workload conditions. Each test
-    // models a simple "plant" where throughput is a function of buffer
-    // size, and asserts that the controller settles within a tolerance
-    // band around the optimal operating point.
-
-    /// Synthetic plant model: throughput = min(k * buffer_size, capacity).
-    ///
-    /// This models the observation that larger buffers reduce syscall
-    /// overhead and improve throughput up to the link's physical limit.
-    fn linear_plant(buffer_size: usize, k: f64, capacity: f64) -> u64 {
-        (k * buffer_size as f64).min(capacity) as u64
-    }
-
-    #[test]
-    fn convergence_constant_workload_stabilizes() {
-        // Under a constant synthetic workload, the controller should
-        // converge to a stable buffer size and stay there. We verify
-        // that the buffer size stops changing (within a small window)
-        // after convergence.
-        let setpoint = 50 * 1024 * 1024; // 50 MB/s
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(2 * 1024 * 1024)
-            .build();
-
-        let mid = (16 * 1024 + 2 * 1024 * 1024) / 2;
-        let k = setpoint as f64 / mid as f64;
-        let cap = setpoint as f64 * 4.0;
-
-        let mut now = Instant::now();
-        // Run 100 samples to let the controller converge.
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // Collect 20 more samples and verify the buffer size is stable.
-        let mut sizes = Vec::with_capacity(20);
-        for _ in 0..20 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-            sizes.push(ctrl.buffer_size());
-        }
-
-        let min_s = *sizes.iter().min().unwrap();
-        let max_s = *sizes.iter().max().unwrap();
-        let range_pct = if min_s > 0 {
-            (max_s - min_s) as f64 / min_s as f64
-        } else {
-            0.0
-        };
-        assert!(
-            range_pct < 0.05,
-            "buffer size should be stable after convergence: min={min_s}, max={max_s}, range={range_pct:.2}%"
-        );
-    }
-
-    #[test]
-    fn convergence_buffer_grows_when_throughput_improves_with_larger_buffers() {
-        // Plant: throughput scales linearly with buffer size up to a cap.
-        // When throughput is below the setpoint and larger buffers help,
-        // the controller should increase the buffer size.
-        let setpoint = 100 * 1024 * 1024; // 100 MB/s
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(4 * 1024 * 1024)
-            .build();
-
-        let initial = ctrl.buffer_size();
-        let mid = (16 * 1024 + 4 * 1024 * 1024) / 2;
-        let k = setpoint as f64 / mid as f64;
-        let cap = setpoint as f64 * 4.0;
-
-        let mut now = Instant::now();
-        // Feed samples where throughput scales with buffer size.
-        // Starting from midpoint, throughput = k * buffer_size < setpoint
-        // because initial buffer = midpoint gives exactly setpoint, but
-        // we start with throughput slightly below setpoint by using half
-        // the buffer size.
-        for _ in 0..30 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let final_size = ctrl.buffer_size();
-        // The controller should have adjusted buffer size. Since the
-        // plant is linear and setpoint matches the midpoint, the controller
-        // should converge near the initial size. The key assertion is that
-        // it explored and settled - not that it moved in a specific
-        // direction from the already-optimal start.
-        let _ = initial;
-        let throughput = linear_plant(final_size, k, cap);
-        let error_pct = (throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.15,
-            "controller should converge near setpoint: throughput={throughput}, setpoint={setpoint}, error={error_pct:.2}"
-        );
-    }
-
-    #[test]
-    fn convergence_buffer_shrinks_when_oversized() {
-        // Plant: throughput peaks at a moderate buffer size and stays flat
-        // beyond that. The controller should not keep growing the buffer
-        // past the saturation point - it should shrink back toward the
-        // minimum effective size.
-        let setpoint = 80 * 1024 * 1024; // 80 MB/s
-        let optimal_buf = 128 * 1024; // throughput saturates at 128 KB
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(2 * 1024 * 1024)
-            .build();
-
-        let k = setpoint as f64 / optimal_buf as f64;
-        let cap = setpoint as f64;
-
-        let mut now = Instant::now();
-        // Run the control loop. Throughput saturates at setpoint regardless
-        // of buffer size above optimal_buf.
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // Once throughput hits the setpoint (error -> 0), the controller
-        // should stop growing. The integral term prevents it from shrinking
-        // below the point where throughput = setpoint.
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.15,
-            "controller should maintain throughput near setpoint: throughput={final_throughput}, setpoint={setpoint}"
-        );
-    }
-
-    #[test]
-    fn convergence_responds_to_sudden_throughput_drop() {
-        // Simulate a workload change: throughput drops suddenly (e.g.,
-        // link degradation). The controller should increase the buffer
-        // to compensate.
-        //
-        // Plant k must satisfy K_p * k < 1 for closed-loop stability.
-        // With k=0.5 and default K_p=0.6, loop gain = 0.3 - well damped.
-        let min = 16 * 1024;
-        let max = 4 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k_fast = 0.5;
-        let setpoint = (k_fast * mid as f64) as u64;
-        let cap_fast = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Phase 1: Converge under normal conditions.
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k_fast, cap_fast);
-            ctrl.observe_at(throughput, now);
-        }
-        let size_before_drop = ctrl.buffer_size();
-
-        // Phase 2: Throughput drops by 50% (slower link).
-        // k_slow=0.25, loop gain = 0.6 * 0.25 = 0.15. Very stable.
-        let k_slow = k_fast * 0.5;
-        let cap_slow = cap_fast * 0.5;
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k_slow, cap_slow);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let size_after_drop = ctrl.buffer_size();
-        assert!(
-            size_after_drop > size_before_drop,
-            "controller should grow buffer after throughput drop: before={size_before_drop}, after={size_after_drop}"
-        );
-    }
-
-    #[test]
-    fn convergence_responds_to_sudden_throughput_increase() {
-        // Simulate throughput improvement: controller should eventually
-        // shrink the buffer since less buffer is needed to hit setpoint.
-        //
-        // k_fast = 1.0 after the 2x gain increase. Gains are reduced
-        // so K_p * k_fast = 0.3 stays within the discrete stability
-        // margin; derivative damping K_d/dt * k = 0.2 prevents ringing.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k_slow = 0.5;
-        let setpoint = (k_slow * mid as f64) as u64;
-        let cap_slow = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Phase 1: Converge under slow conditions.
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k_slow, cap_slow);
-            ctrl.observe_at(throughput, now);
-        }
-        let size_before_increase = ctrl.buffer_size();
-
-        // Phase 2: Link gets 2x faster - throughput exceeds setpoint
-        // at the current buffer size, so the controller should shrink.
-        let k_fast = k_slow * 2.0;
-        let cap_fast = cap_slow * 2.0;
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k_fast, cap_fast);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let size_after_increase = ctrl.buffer_size();
-        assert!(
-            size_after_increase < size_before_increase,
-            "controller should shrink buffer after throughput increase: before={size_before_increase}, after={size_after_increase}"
-        );
-    }
-
-    #[test]
-    fn convergence_minimum_size_enforced_under_extreme_overshoot() {
-        // Even when throughput massively exceeds the setpoint (error
-        // is deeply negative), the buffer size must never drop below
-        // the configured minimum.
-        let min = 32 * 1024;
-        let ctrl = ControllerConfig::new(1_000_000) // 1 MB/s setpoint
-            .min_size(min)
-            .max_size(1024 * 1024)
-            .build();
-
-        let mut now = Instant::now();
-        // Feed extremely high throughput (1 GB/s - 1000x the setpoint).
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(1_000_000_000, now);
-        }
-
-        assert_eq!(ctrl.buffer_size(), min, "buffer size must clamp at minimum");
-    }
-
-    #[test]
-    fn convergence_maximum_size_enforced_under_extreme_undershoot() {
-        // Even when throughput is zero (error equals the full setpoint),
-        // the buffer size must never exceed the configured maximum.
-        let max = 512 * 1024;
-        let ctrl = ControllerConfig::new(lan_setpoint())
-            .min_size(16 * 1024)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(0, now);
-        }
-
-        assert_eq!(ctrl.buffer_size(), max, "buffer size must clamp at maximum");
-    }
-
-    #[test]
-    fn convergence_oscillation_dampens_over_time() {
-        // Feed alternating high/low throughput samples. The derivative
-        // term should damp oscillations so the buffer size variance
-        // decreases over time (later samples have smaller swings).
-        let setpoint = 50 * 1024 * 1024;
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(2 * 1024 * 1024)
-            .build();
-
-        let mut now = Instant::now();
-        let low = setpoint / 4;
-        let high = setpoint * 4;
-
-        // Phase 1: first 20 alternating samples.
-        let mut early_sizes = Vec::with_capacity(20);
-        for i in 0..20 {
-            now += Duration::from_millis(100);
-            let throughput = if i % 2 == 0 { low } else { high };
-            ctrl.observe_at(throughput, now);
-            early_sizes.push(ctrl.buffer_size());
-        }
-
-        // Phase 2: next 20 alternating samples.
-        let mut late_sizes = Vec::with_capacity(20);
-        for i in 0..20 {
-            now += Duration::from_millis(100);
-            let throughput = if i % 2 == 0 { low } else { high };
-            ctrl.observe_at(throughput, now);
-            late_sizes.push(ctrl.buffer_size());
-        }
-
-        let early_range = early_sizes.iter().max().unwrap() - early_sizes.iter().min().unwrap();
-        let late_range = late_sizes.iter().max().unwrap() - late_sizes.iter().min().unwrap();
-
-        // The late window should have equal or smaller oscillation amplitude.
-        assert!(
-            late_range <= early_range,
-            "oscillation should dampen: early_range={early_range}, late_range={late_range}"
-        );
-    }
-
-    #[test]
-    fn convergence_step_change_settles_within_20_samples() {
-        // After a step change in the plant's gain (simulating a new
-        // workload), the controller should re-converge within 20
-        // samples (2 seconds at 100 ms intervals).
-        //
-        // k1=0.5, k2=1.0. Gains scaled so K_p*k2 = 0.3 for stability.
-        let min = 16 * 1024;
-        let max = 4 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k1 = 0.5;
-        let setpoint = (k1 * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        // Converge under k1.
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k1, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // Step change: k doubles (half the buffer needed for same throughput).
-        let k2 = k1 * 2.0;
-        let mut converged_at = None;
-        for i in 0..20 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k2, cap);
-            ctrl.observe_at(throughput, now);
-            let measured = linear_plant(ctrl.buffer_size(), k2, cap);
-            let error_pct = (measured as f64 - setpoint as f64).abs() / setpoint as f64;
-            if error_pct < 0.10 {
-                converged_at = Some(i);
-                break;
-            }
-        }
-
-        assert!(
-            converged_at.is_some(),
-            "controller should re-converge within 20 samples after step change (final buffer={}, throughput={})",
-            ctrl.buffer_size(),
-            linear_plant(ctrl.buffer_size(), k2, cap)
-        );
-    }
-
-    #[test]
-    fn convergence_pure_integral_eliminates_offset() {
-        // With K_p = 0 and K_d = 0, a pure integral controller should
-        // still drive the steady-state error to zero given enough samples.
-        let setpoint = 10_000_000; // 10 MB/s
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.0, 0.3, 0.0)
-            .min_size(4 * 1024)
-            .max_size(1024 * 1024)
-            .build();
-
-        let mid = (4 * 1024 + 1024 * 1024) / 2;
-        let k = setpoint as f64 / mid as f64;
-        let cap = setpoint as f64 * 4.0;
-
-        let mut now = Instant::now();
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.10,
-            "pure integral should eliminate steady-state error: throughput={final_throughput}, setpoint={setpoint}, error={error_pct:.2}"
-        );
-    }
-
-    #[test]
-    fn convergence_wan_preset_with_high_jitter() {
-        // WAN-like conditions: lower setpoint, higher jitter. The
-        // controller should still converge despite noisy measurements.
-        let min = 4 * 1024;
-        let max = 512 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.4, 0.15, 0.03)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let cap = setpoint as f64 * 4.0;
-
-        let mut now = Instant::now();
-        // Simulate jittery throughput: base + noise pattern.
-        let noise_pattern: [f64; 8] = [0.7, 1.3, 0.9, 1.1, 0.8, 1.2, 0.95, 1.05];
-        for i in 0..100 {
-            now += Duration::from_millis(100);
-            let base_throughput = linear_plant(ctrl.buffer_size(), k, cap) as f64;
-            let jittered = (base_throughput * noise_pattern[i % noise_pattern.len()]) as u64;
-            ctrl.observe_at(jittered, now);
-        }
-
-        // Despite jitter, throughput should be near the setpoint.
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.25,
-            "WAN preset should converge despite jitter: throughput={final_throughput}, setpoint={setpoint}, error={error_pct:.2}"
-        );
-    }
-
-    // --- Extended convergence tests (task #2096) ---
-    //
-    // These tests verify convergence under more complex workload patterns,
-    // boundary conditions, and multi-phase scenarios that exercise the
-    // controller's stability beyond the basic single-pattern tests above.
-
-    /// Variance of a slice of `usize` values, returned as f64.
-    fn variance(values: &[usize]) -> f64 {
-        if values.is_empty() {
-            return 0.0;
-        }
-        let mean = values.iter().sum::<usize>() as f64 / values.len() as f64;
-        values
-            .iter()
-            .map(|&v| (v as f64 - mean).powi(2))
-            .sum::<f64>()
-            / values.len() as f64
-    }
-
-    #[test]
-    fn convergence_multi_phase_ramp_plateau_decay() {
-        // Workload pattern: ramp up throughput linearly, hold at plateau,
-        // then decay. The controller must track each phase and stabilize.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Phase 1: Ramp up from 10% to 100% of plant capacity over 50 samples.
-        for i in 0..50 {
-            now += Duration::from_millis(100);
-            let ramp_factor = 0.1 + 0.9 * (i as f64 / 49.0);
-            let throughput = linear_plant(ctrl.buffer_size(), k * ramp_factor, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // Phase 2: Plateau at full plant capacity for 50 samples.
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // Verify plateau stability.
-        let mut plateau_sizes = Vec::with_capacity(20);
-        for _ in 0..20 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-            plateau_sizes.push(ctrl.buffer_size());
-        }
-        let plateau_var = variance(&plateau_sizes);
-        let plateau_mean = plateau_sizes.iter().sum::<usize>() as f64 / plateau_sizes.len() as f64;
-        let cv = if plateau_mean > 0.0 {
-            plateau_var.sqrt() / plateau_mean
-        } else {
-            0.0
-        };
-        assert!(
-            cv < 0.10,
-            "plateau phase should be stable: cv={cv:.4}, mean={plateau_mean:.0}"
-        );
-
-        // Phase 3: Decay plant gain by 50% and let the controller adapt.
-        let k_slow = k * 0.5;
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k_slow, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // After decay, throughput should still be near setpoint (controller
-        // grew the buffer to compensate for reduced plant gain).
-        let final_throughput = linear_plant(ctrl.buffer_size(), k_slow, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.20,
-            "controller should track decay: throughput={final_throughput}, setpoint={setpoint}, error={error_pct:.2}"
-        );
-    }
-
-    #[test]
-    fn convergence_zero_load_sustained_no_pathological_behavior() {
-        // Feed zero throughput for many iterations. The controller should
-        // saturate at max_size without panicking, producing NaN, or any
-        // other pathological behavior.
-        let max = 512 * 1024;
-        let ctrl = ControllerConfig::new(50 * 1024 * 1024)
-            .min_size(16 * 1024)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..500 {
-            now += Duration::from_millis(100);
-            let size = ctrl.observe_at(0, now);
-            assert!(size >= ctrl.min_size(), "size below min: {size}");
-            assert!(size <= ctrl.max_size(), "size above max: {size}");
-        }
-
-        // Should be pinned at max after sustained zero throughput.
-        assert_eq!(ctrl.buffer_size(), max);
-
-        // Verify the integrator is clamped (not infinite).
-        let state = ctrl.state.lock().unwrap();
-        assert!(state.integral.is_finite(), "integral must be finite");
-    }
-
-    #[test]
-    fn convergence_max_throughput_sustained_no_pathological_behavior() {
-        // Feed extremely high throughput (u64::MAX / 2) for many iterations.
-        // The controller should saturate at min_size without overflow.
-        let min = 16 * 1024;
-        let ctrl = ControllerConfig::new(1_000_000)
-            .min_size(min)
-            .max_size(1024 * 1024)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..500 {
-            now += Duration::from_millis(100);
-            let size = ctrl.observe_at(u64::MAX / 2, now);
-            assert!(size >= min, "size below min: {size}");
-            assert!(size <= ctrl.max_size(), "size above max: {size}");
-        }
-
-        assert_eq!(ctrl.buffer_size(), min);
-
-        let state = ctrl.state.lock().unwrap();
-        assert!(state.integral.is_finite(), "integral must be finite");
-        assert!(state.prev_error.is_finite(), "prev_error must be finite");
-    }
-
-    #[test]
-    fn convergence_sawtooth_load_bounded_oscillation() {
-        // Feed a repeating sawtooth pattern: throughput linearly ramps up
-        // then drops to zero, repeating. The controller's oscillation
-        // amplitude should remain bounded and not grow over time.
-        let setpoint = 50 * 1024 * 1024u64;
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(16 * 1024)
-            .max_size(2 * 1024 * 1024)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Run 5 sawtooth cycles, 20 samples each.
-        let mut cycle_ranges = Vec::new();
-        for _cycle in 0..5 {
-            let mut sizes = Vec::with_capacity(20);
-            for i in 0..20 {
-                now += Duration::from_millis(100);
-                // Ramp from 0 to 2x setpoint linearly.
-                let fraction = i as f64 / 19.0;
-                let throughput = (setpoint as f64 * 2.0 * fraction) as u64;
-                ctrl.observe_at(throughput, now);
-                sizes.push(ctrl.buffer_size());
-            }
-            let cycle_min = *sizes.iter().min().unwrap();
-            let cycle_max = *sizes.iter().max().unwrap();
-            cycle_ranges.push(cycle_max - cycle_min);
-        }
-
-        // The oscillation range should not grow over successive cycles.
-        // Compare the last cycle's range to the first cycle's range.
-        let first_range = cycle_ranges[0];
-        let last_range = *cycle_ranges.last().unwrap();
-        assert!(
-            last_range <= first_range + first_range / 10,
-            "oscillation should not grow: first_range={first_range}, last_range={last_range}"
-        );
-    }
-
-    #[test]
-    fn convergence_reset_and_reconverge() {
-        // Converge the controller, then reset the PID accumulators and
-        // verify it re-converges to the same operating point.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Converge initially.
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-        let converged_size = ctrl.buffer_size();
-        let converged_throughput = linear_plant(converged_size, k, cap);
-        let error_before = (converged_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-
-        // Reset the PID state.
-        ctrl.reset();
-
-        // Re-converge.
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-        let reconverged_size = ctrl.buffer_size();
-        let reconverged_throughput = linear_plant(reconverged_size, k, cap);
-        let error_after = (reconverged_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-
-        // Both convergences should reach similar accuracy.
-        assert!(
-            error_before < 0.15,
-            "initial convergence failed: error={error_before:.2}"
-        );
-        assert!(
-            error_after < 0.15,
-            "post-reset reconvergence failed: error={error_after:.2}"
-        );
-    }
-
-    #[test]
-    fn convergence_narrow_min_max_window() {
-        // When min_size and max_size are very close together, the controller
-        // should still converge without oscillating between the two bounds.
-        let min = 100 * 1024;
-        let max = 120 * 1024;
-        let setpoint = 50 * 1024 * 1024u64;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(setpoint, now);
-        }
-
-        // With throughput exactly at setpoint, the buffer should stabilize.
-        let mut sizes = Vec::with_capacity(20);
-        for _ in 0..20 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(setpoint, now);
-            sizes.push(ctrl.buffer_size());
-        }
-
-        // All sizes should be the same (steady state with zero error).
-        let unique: std::collections::HashSet<_> = sizes.iter().collect();
-        assert!(
-            unique.len() <= 2,
-            "narrow window should produce stable output, got {} distinct sizes: {:?}",
-            unique.len(),
-            sizes
-        );
-    }
-
-    #[test]
-    fn convergence_min_equals_max_stays_fixed() {
-        // When min_size == max_size, the controller has no room to adjust.
-        // Buffer size must always equal that value.
-        let fixed = 128 * 1024;
-        let ctrl = ControllerConfig::new(50 * 1024 * 1024)
-            .min_size(fixed)
-            .max_size(fixed)
-            .build();
-
-        let mut now = Instant::now();
-        for sample in [0u64, 1, 1_000_000, 100_000_000, u64::MAX / 2] {
-            now += Duration::from_millis(100);
-            let size = ctrl.observe_at(sample, now);
-            assert_eq!(
-                size, fixed,
-                "size must equal fixed bound regardless of input"
-            );
-        }
-    }
-
-    #[test]
-    fn convergence_speed_proportional_only() {
-        // A proportional-only controller (no I, no D) should converge
-        // within a bounded number of iterations for a linear plant.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.3;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.5, 0.0, 0.0)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        let mut converged_at = None;
-        for i in 0..100 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-            let measured = linear_plant(ctrl.buffer_size(), k, cap);
-            let error_pct = (measured as f64 - setpoint as f64).abs() / setpoint as f64;
-            if error_pct < 0.05 && converged_at.is_none() {
-                converged_at = Some(i);
-            }
-        }
-
-        // P-only should converge quickly (proportional acts immediately)
-        // but may have residual steady-state error. We just check it
-        // reaches within 5% at some point.
-        assert!(
-            converged_at.is_some(),
-            "P-only controller should converge within 100 samples"
-        );
-        assert!(
-            converged_at.unwrap() < 30,
-            "P-only convergence should be fast, but took {} samples",
-            converged_at.unwrap()
-        );
-    }
-
-    #[test]
-    fn convergence_speed_full_pid_faster_than_integral_only() {
-        // Full PID should converge at least as fast as pure I controller
-        // for the same plant model.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.3;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let t0 = Instant::now();
-
-        let ctrl_pid = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-        let ctrl_i_only = ControllerConfig::new(setpoint)
-            .gains(0.0, 0.1, 0.0)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let tolerance = 0.10;
-        let max_samples = 200;
-
-        let find_convergence = |ctrl: &AdaptiveBufferController, start: Instant| -> Option<usize> {
-            let mut t = start;
-            for i in 0..max_samples {
-                t += Duration::from_millis(100);
-                let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-                ctrl.observe_at(throughput, t);
-                let measured = linear_plant(ctrl.buffer_size(), k, cap);
-                let error_pct = (measured as f64 - setpoint as f64).abs() / setpoint as f64;
-                if error_pct < tolerance {
-                    return Some(i);
-                }
-            }
-            None
-        };
-
-        let pid_samples = find_convergence(&ctrl_pid, t0);
-        let i_samples = find_convergence(&ctrl_i_only, t0);
-
-        assert!(
-            pid_samples.is_some(),
-            "PID controller should converge within {max_samples} samples"
-        );
-        assert!(
-            i_samples.is_some(),
-            "I-only controller should converge within {max_samples} samples"
-        );
-        assert!(
-            pid_samples.unwrap() <= i_samples.unwrap(),
-            "PID should converge at least as fast as I-only: PID={}, I-only={}",
-            pid_samples.unwrap(),
-            i_samples.unwrap()
-        );
-    }
-
-    #[test]
-    fn convergence_derivative_kick_suppression() {
-        // When the setpoint changes (simulated by changing the plant gain),
-        // the derivative term acts on the error derivative, not the process
-        // variable derivative. A sudden error spike should not cause the
-        // controller to overshoot wildly. We compare overshoot with D=0
-        // vs D>0.
-        let min = 16 * 1024;
-        let max = 4 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-        let t0 = Instant::now();
-
-        let ctrl_no_d = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.0)
-            .min_size(min)
-            .max_size(max)
-            .build();
-        let ctrl_with_d = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.05)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        // Converge both under normal conditions.
-        let mut t = t0;
-        for _ in 0..80 {
-            t += Duration::from_millis(100);
-            let tp_no_d = linear_plant(ctrl_no_d.buffer_size(), k, cap);
-            ctrl_no_d.observe_at(tp_no_d, t);
-            let tp_with_d = linear_plant(ctrl_with_d.buffer_size(), k, cap);
-            ctrl_with_d.observe_at(tp_with_d, t);
-        }
-
-        let steady_no_d = ctrl_no_d.buffer_size();
-        let steady_with_d = ctrl_with_d.buffer_size();
-
-        // Sudden plant change: throughput drops to 10% of before.
-        let k_slow = k * 0.1;
-        t += Duration::from_millis(100);
-        let tp_no_d = linear_plant(ctrl_no_d.buffer_size(), k_slow, cap);
-        ctrl_no_d.observe_at(tp_no_d, t);
-        let tp_with_d = linear_plant(ctrl_with_d.buffer_size(), k_slow, cap);
-        ctrl_with_d.observe_at(tp_with_d, t);
-
-        let kick_no_d = (ctrl_no_d.buffer_size() as i64 - steady_no_d as i64).unsigned_abs();
-        let kick_with_d = (ctrl_with_d.buffer_size() as i64 - steady_with_d as i64).unsigned_abs();
-
-        // The derivative-equipped controller should have a larger or equal
-        // initial response (D adds to the proportional kick for same-sign
-        // derivative), but the key property is that neither produces an
-        // output outside the configured bounds.
-        assert!(
-            ctrl_no_d.buffer_size() >= min && ctrl_no_d.buffer_size() <= max,
-            "no-D controller output out of bounds: {}",
-            ctrl_no_d.buffer_size()
-        );
-        assert!(
-            ctrl_with_d.buffer_size() >= min && ctrl_with_d.buffer_size() <= max,
-            "with-D controller output out of bounds: {}",
-            ctrl_with_d.buffer_size()
-        );
-
-        // Both should have responded (non-zero kick).
-        assert!(
-            kick_no_d > 0 || kick_with_d > 0,
-            "at least one controller should react to step change"
-        );
-    }
-
-    #[test]
-    fn convergence_rapid_sample_intervals() {
-        // Feed samples at 1 ms intervals (10x faster than default). The
-        // controller should still converge due to the MIN_DT clamp.
-        let min = 16 * 1024;
-        let max = 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .sample_interval_ms(1)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..500 {
-            now += Duration::from_millis(1);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.15,
-            "should converge with rapid samples: error={error_pct:.2}"
-        );
-    }
-
-    #[test]
-    fn convergence_slow_sample_intervals() {
-        // Feed samples at 5 s intervals (max dt). The MAX_DT clamp
-        // prevents integral windup.
-        let min = 16 * 1024;
-        let max = 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..50 {
-            now += Duration::from_secs(5);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.20,
-            "should converge with slow samples: error={error_pct:.2}"
-        );
-
-        // Verify integral did not wind up to infinity.
-        let state = ctrl.state.lock().unwrap();
-        assert!(
-            state.integral.is_finite(),
-            "integral must be finite under slow sampling"
-        );
-    }
-
-    #[test]
-    fn convergence_stall_then_resume() {
-        // Simulate a producer stall (large gap between samples, then
-        // normal sampling resumes). The controller should recover.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Converge normally.
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-        let size_before_stall = ctrl.buffer_size();
-
-        // Stall for 30 seconds (well beyond MAX_DT of 5s).
-        now += Duration::from_secs(30);
-        let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        ctrl.observe_at(throughput, now);
-
-        // Size should still be within bounds.
-        assert!(ctrl.buffer_size() >= min);
-        assert!(ctrl.buffer_size() <= max);
-
-        // Resume normal sampling and verify reconvergence.
-        for _ in 0..80 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.15,
-            "should reconverge after stall: before_stall={size_before_stall}, after={}, error={error_pct:.2}",
-            ctrl.buffer_size()
-        );
-    }
-
-    #[test]
-    fn convergence_alternating_fast_and_slow_plant() {
-        // Alternate between a fast plant (throughput above setpoint) and a
-        // slow plant (throughput below setpoint). The controller should
-        // remain bounded and eventually stabilize (variance decreases or
-        // stays bounded across cycles).
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.5;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // 10 cycles: 10 samples with k*1.5 (above setpoint), then 10
-        // with k*0.7 (below setpoint). The average throughput at midpoint
-        // is ~1.1*setpoint, so the controller should settle to a moderate
-        // buffer size.
-        let k_fast = k * 1.5;
-        let k_slow = k * 0.7;
-
-        // Early window: collect sizes from first 5 cycles.
-        let mut early_sizes = Vec::new();
-        for _ in 0..5 {
-            for _ in 0..10 {
-                now += Duration::from_millis(100);
-                let throughput = linear_plant(ctrl.buffer_size(), k_fast, cap);
-                ctrl.observe_at(throughput, now);
-                early_sizes.push(ctrl.buffer_size());
-            }
-            for _ in 0..10 {
-                now += Duration::from_millis(100);
-                let throughput = linear_plant(ctrl.buffer_size(), k_slow, cap);
-                ctrl.observe_at(throughput, now);
-                early_sizes.push(ctrl.buffer_size());
-            }
-        }
-
-        // Late window: collect sizes from next 5 cycles.
-        let mut late_sizes = Vec::new();
-        for _ in 0..5 {
-            for _ in 0..10 {
-                now += Duration::from_millis(100);
-                let throughput = linear_plant(ctrl.buffer_size(), k_fast, cap);
-                ctrl.observe_at(throughput, now);
-                late_sizes.push(ctrl.buffer_size());
-            }
-            for _ in 0..10 {
-                now += Duration::from_millis(100);
-                let throughput = linear_plant(ctrl.buffer_size(), k_slow, cap);
-                ctrl.observe_at(throughput, now);
-                late_sizes.push(ctrl.buffer_size());
-            }
-        }
-
-        // Late variance should be bounded (not growing).
-        let early_var = variance(&early_sizes);
-        let late_var = variance(&late_sizes);
-        assert!(
-            late_var <= early_var * 1.5 + 1.0,
-            "oscillation should not grow: early_var={early_var:.0}, late_var={late_var:.0}"
-        );
-
-        // All sizes must be within bounds.
-        for &s in early_sizes.iter().chain(late_sizes.iter()) {
-            assert!(
-                s >= min && s <= max,
-                "size {s} out of bounds [{min}, {max}]"
-            );
-        }
-    }
-
-    #[test]
-    fn convergence_monotonic_approach_from_min() {
-        // Start the controller at min_size (by setting min close to initial)
-        // and verify it monotonically approaches the setpoint from below
-        // (no overshoots that dip below the starting point). Use a plant
-        // where setpoint is achievable well within the buffer range.
-        let min = 16 * 1024;
-        let max = 4 * 1024 * 1024;
-        let mid = (min + max) / 2;
-
-        // Set gains low enough that loop gain K_p * k < 1 for stability.
-        let k = 0.3;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-
-        // Use min_size close to the default initial midpoint so the
-        // controller starts from a consistent position.
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.2, 0.05, 0.01)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let initial = ctrl.buffer_size();
-        let mut now = Instant::now();
-
-        // Feed low throughput so the controller grows the buffer.
-        let mut floor = initial;
-        let mut monotonic_violations = 0;
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            // Plant returns less than setpoint so error is positive -> grow.
-            let throughput = linear_plant(ctrl.buffer_size(), k * 0.5, cap);
-            ctrl.observe_at(throughput, now);
-            let current = ctrl.buffer_size();
-            if current < floor {
-                monotonic_violations += 1;
-            }
-            floor = floor.max(current);
-        }
-
-        // Allow at most a couple of minor dips due to derivative kick
-        // when the error rate changes sign.
-        assert!(
-            monotonic_violations <= 3,
-            "growth from below setpoint should be mostly monotonic: {monotonic_violations} violations",
-        );
-    }
-
-    #[test]
-    fn convergence_high_noise_amplitude_bounded_output() {
-        // Even under extremely noisy measurements (noise amplitude equals
-        // the setpoint), the controller must keep its output within bounds.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let setpoint = 50 * 1024 * 1024u64;
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-
-        // Noise pattern with full-range swings around the setpoint.
-        let noise_factors: [f64; 10] = [0.0, 2.0, 0.1, 1.9, 0.3, 1.7, 0.5, 1.5, 0.2, 1.8];
-        for i in 0..200 {
-            now += Duration::from_millis(100);
-            let throughput = (setpoint as f64 * noise_factors[i % noise_factors.len()]) as u64;
-            let size = ctrl.observe_at(throughput, now);
-            assert!(
-                size >= min,
-                "output below min at iteration {i}: {size} < {min}"
-            );
-            assert!(
-                size <= max,
-                "output above max at iteration {i}: {size} > {max}"
-            );
-        }
-    }
-
-    #[test]
-    fn convergence_setpoint_at_plant_maximum() {
-        // When the setpoint equals the plant's maximum output, the
-        // controller should grow the buffer to the point where the
-        // plant saturates and then stabilize at max buffer size.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        // Choose k so that throughput = setpoint at the midpoint.
-        // The plant cap equals setpoint, so at mid: k*mid = setpoint.
-        // For buffer sizes > mid, throughput saturates at plant_cap.
-        let setpoint = 50 * 1024 * 1024u64;
-        let k = setpoint as f64 / mid as f64;
-        let plant_cap = setpoint as f64; // Plant saturates at setpoint.
-
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.3, 0.1, 0.02)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut now = Instant::now();
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            let throughput = linear_plant(ctrl.buffer_size(), k, plant_cap);
-            ctrl.observe_at(throughput, now);
-        }
-
-        // At the plant ceiling, throughput saturates. The controller's
-        // error goes to zero once throughput matches setpoint. The buffer
-        // should stabilize (not keep growing) once throughput = setpoint.
-        let final_throughput = linear_plant(ctrl.buffer_size(), k, plant_cap);
-        let error_pct = (final_throughput as f64 - setpoint as f64).abs() / setpoint as f64;
-        assert!(
-            error_pct < 0.05,
-            "throughput should match setpoint at saturation: throughput={final_throughput}, setpoint={setpoint}"
-        );
-    }
-
-    #[test]
-    fn convergence_asymmetric_gains_bias() {
-        // With high P and low I, convergence should be fast but have
-        // residual steady-state offset. With high I and low P, convergence
-        // should be slower but eliminate offset. This verifies the relative
-        // behavior matches PID theory.
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-        let mid = (min + max) / 2;
-        let k = 0.3;
-        let setpoint = (k * mid as f64) as u64;
-        let cap = setpoint as f64 * 4.0;
-        let t0 = Instant::now();
-
-        let ctrl_high_p = ControllerConfig::new(setpoint)
-            .gains(0.5, 0.01, 0.0)
-            .min_size(min)
-            .max_size(max)
-            .build();
-        let ctrl_high_i = ControllerConfig::new(setpoint)
-            .gains(0.05, 0.3, 0.0)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        let mut t = t0;
-        for _ in 0..300 {
-            t += Duration::from_millis(100);
-            let tp_hp = linear_plant(ctrl_high_p.buffer_size(), k, cap);
-            ctrl_high_p.observe_at(tp_hp, t);
-            let tp_hi = linear_plant(ctrl_high_i.buffer_size(), k, cap);
-            ctrl_high_i.observe_at(tp_hi, t);
-        }
-
-        let final_hp = linear_plant(ctrl_high_p.buffer_size(), k, cap);
-        let final_hi = linear_plant(ctrl_high_i.buffer_size(), k, cap);
-        let error_hp = (final_hp as f64 - setpoint as f64).abs() / setpoint as f64;
-        let error_hi = (final_hi as f64 - setpoint as f64).abs() / setpoint as f64;
-
-        // High-I should have lower steady-state error than high-P.
-        assert!(
-            error_hi <= error_hp + 0.05,
-            "high-I should have lower or equal steady-state error: error_hi={error_hi:.3}, error_hp={error_hp:.3}"
-        );
-    }
-
-    #[test]
-    fn convergence_concurrent_observers() {
-        // Multiple threads calling observe_at concurrently should not cause
-        // panics, data corruption, or outputs outside bounds.
-        use std::sync::Arc;
-        use std::thread;
-
-        let setpoint = 50 * 1024 * 1024u64;
-        let min = 16 * 1024;
-        let max = 2 * 1024 * 1024;
-
-        let ctrl = Arc::new(
-            ControllerConfig::new(setpoint)
-                .min_size(min)
-                .max_size(max)
-                .build(),
-        );
-
-        let handles: Vec<_> = (0..4)
-            .map(|_| {
-                let ctrl = Arc::clone(&ctrl);
-                thread::spawn(move || {
-                    for _ in 0..200 {
-                        let size = ctrl.observe(setpoint);
-                        assert!(size >= min, "concurrent output below min: {size}");
-                        assert!(size <= max, "concurrent output above max: {size}");
-                    }
-                })
-            })
-            .collect();
-
-        for h in handles {
-            h.join().expect("thread panicked during concurrent observe");
-        }
-    }
-
-    // --- Property-style convergence tests (task #2096) ---
-    //
-    // These tests feed the controller open-loop sample streams (no plant
-    // feedback) and assert three closed-loop properties:
-    //   1. steady-state convergence under a constant input;
-    //   2. bounded output amplitude under a noisy input;
-    //   3. respect for the configured upper buffer-size cap under a
-    //      saturating signal that would otherwise wind the integrator
-    //      unbounded.
-    //
-    // All randomness is driven by a deterministic seeded SplitMix64 PRNG so
-    // the tests are reproducible across runs and platforms. Tolerances are
-    // intentionally loose since these are convergence-property checks, not
-    // exact-value checks.
-
-    /// Deterministic 64-bit SplitMix64 PRNG state. Reseeded per-test from a
-    /// fixed constant so the resulting sample stream is identical on every
-    /// invocation, regardless of platform RNG behaviour.
-    struct SplitMix64(u64);
-
-    impl SplitMix64 {
-        fn new(seed: u64) -> Self {
-            Self(seed)
-        }
-
-        fn next_u64(&mut self) -> u64 {
-            self.0 = self.0.wrapping_add(0x9E37_79B9_7F4A_7C15);
-            let mut z = self.0;
-            z = (z ^ (z >> 30)).wrapping_mul(0xBF58_476D_1CE4_E5B9);
-            z = (z ^ (z >> 27)).wrapping_mul(0x94D0_49BB_1331_11EB);
-            z ^ (z >> 31)
-        }
-
-        /// Uniform in `[0.0, 1.0)`.
-        fn next_unit(&mut self) -> f64 {
-            // 53 bits of precision; standard f64 unit sample.
-            (self.next_u64() >> 11) as f64 / (1u64 << 53) as f64
-        }
-
-        /// Approximate standard-normal sample via the Box-Muller transform.
-        fn next_normal(&mut self) -> f64 {
-            // Guard against log(0) by clamping the uniform draw away from 0.
-            let u1 = self.next_unit().max(f64::MIN_POSITIVE);
-            let u2 = self.next_unit();
-            (-2.0 * u1.ln()).sqrt() * (std::f64::consts::TAU * u2).cos()
-        }
-    }
-
-    /// Property 1: under a constant input signal, the buffer size converges
-    /// and the residual oscillation amplitude is small (within +/-5% of the
-    /// mean across the post-warm-up window).
-    #[test]
-    fn property_steady_state_convergence_under_constant_input() {
-        // Seed kept here for reproducibility even though this test does not
-        // consume any random samples - the seeded RNG is documented as part
-        // of the property-test suite contract.
-        let _rng = SplitMix64::new(0xCAFE);
-
-        let setpoint = 50 * 1024 * 1024u64;
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.6, 0.2, 0.05)
-            .min_size(16 * 1024)
-            .max_size(2 * 1024 * 1024)
-            .build();
-
-        // Feed a constant throughput equal to the setpoint for 100 samples.
-        // With zero error each step, P and D contribute nothing, and the
-        // integrator stays at whatever steady-state value it converged to.
-        let mut now = Instant::now();
-        for _ in 0..100 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(setpoint, now);
-        }
-
-        // Collect 50 post-warm-up samples and check the oscillation band.
-        let mut sizes = Vec::with_capacity(50);
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            ctrl.observe_at(setpoint, now);
-            sizes.push(ctrl.buffer_size());
-        }
-
-        let mean = sizes.iter().sum::<usize>() as f64 / sizes.len() as f64;
-        let min_s = *sizes.iter().min().unwrap() as f64;
-        let max_s = *sizes.iter().max().unwrap() as f64;
-        // Tolerance: amplitude within +/-5% of the mean. Loose by design;
-        // the property is "stays put", not "exact value".
-        let amplitude = (max_s - min_s) / mean.max(1.0);
-        assert!(
-            amplitude < 0.05,
-            "steady-state amplitude {amplitude:.4} exceeded 5% of mean {mean:.0}"
-        );
-    }
-
-    /// Property 2: under a noisy input (Gaussian, mean below setpoint with
-    /// non-trivial sigma), the controller's output stays within a bounded
-    /// amplitude band - it does not diverge or oscillate without limit.
-    #[test]
-    fn property_noisy_signal_bounded_output() {
-        let mut rng = SplitMix64::new(0xCAFE);
-
-        let setpoint = 100 * 1024 * 1024u64;
-        let min = 16 * 1024;
-        let max = 4 * 1024 * 1024;
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.6, 0.2, 0.05)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        // Signal model: throughput ~ N(mean=0.3*setpoint, sigma=0.1*setpoint).
-        // Mean 0.3 + sigma 0.1 keeps the bulk of samples well below the
-        // setpoint, exercising the growth side of the controller under noise.
-        let mean = 0.3 * setpoint as f64;
-        let sigma = 0.1 * setpoint as f64;
-
-        let mut now = Instant::now();
-        // Warm-up: 50 samples to let the controller settle near its
-        // working point under the noisy stream.
-        for _ in 0..50 {
-            now += Duration::from_millis(100);
-            let raw = mean + sigma * rng.next_normal();
-            let sample = raw.max(0.0) as u64;
-            ctrl.observe_at(sample, now);
-        }
-
-        // Measure: 200 samples post-warm-up.
-        let mut sizes = Vec::with_capacity(200);
-        for _ in 0..200 {
-            now += Duration::from_millis(100);
-            let raw = mean + sigma * rng.next_normal();
-            let sample = raw.max(0.0) as u64;
-            ctrl.observe_at(sample, now);
-            sizes.push(ctrl.buffer_size());
-        }
-
-        let mean_size = sizes.iter().sum::<usize>() as f64 / sizes.len() as f64;
-        let min_s = *sizes.iter().min().unwrap() as f64;
-        let max_s = *sizes.iter().max().unwrap() as f64;
-
-        // Property: output band stays within [0.5x, 2x] of the mean - i.e.
-        // the controller damps the noise rather than amplifying it. This is
-        // a loose bound chosen so genuine divergence (orders of magnitude
-        // swings or rail-to-rail bouncing) is caught while normal residual
-        // jitter from a 10%-sigma signal is accepted.
-        assert!(
-            min_s >= 0.5 * mean_size,
-            "output min {min_s:.0} below 0.5x mean {mean_size:.0}"
-        );
-        assert!(
-            max_s <= 2.0 * mean_size,
-            "output max {max_s:.0} above 2x mean {mean_size:.0}"
-        );
-
-        // And of course, every individual sample respects the hard cap.
-        for &s in &sizes {
-            assert!(s >= min, "output {s} below configured min {min}");
-            assert!(s <= max, "output {s} above configured max {max}");
-        }
-    }
-
-    /// Property 3: under a saturating signal that drives the error term
-    /// hard in one direction for many iterations, the controller respects
-    /// the configured upper bound - anti-windup prevents the integrator
-    /// from pushing the recommended size past `max_size`.
-    #[test]
-    fn property_cap_respected_under_saturating_signal() {
-        // Seed retained for reproducibility parity with the other
-        // property tests; this test is deterministic by construction.
-        let _rng = SplitMix64::new(0xCAFE);
-
-        let min = 16 * 1024;
-        let max = 256 * 1024;
-        let setpoint = 100 * 1024 * 1024u64;
-        let ctrl = ControllerConfig::new(setpoint)
-            .gains(0.6, 0.2, 0.05)
-            .min_size(min)
-            .max_size(max)
-            .build();
-
-        // Feed zero throughput for 500 iterations. Naively, K_i * error * dt
-        // accumulated for 500 * 100ms = 50s would push the integrator to
-        // K_i * setpoint * 50 = 0.2 * 100MB * 50 = 1 GB, far past `max`.
-        // Anti-windup must clamp this so the output never exceeds `max`.
-        let mut now = Instant::now();
-        for i in 0..500 {
-            now += Duration::from_millis(100);
-            let size = ctrl.observe_at(0, now);
-            assert!(
-                size <= max,
-                "iteration {i}: output {size} exceeded configured max {max}"
-            );
-            assert!(
-                size >= min,
-                "iteration {i}: output {size} below configured min {min}"
-            );
-        }
-
-        // After the saturating run, the controller should be pinned at the
-        // upper bound and the integrator must remain finite.
-        assert_eq!(ctrl.buffer_size(), max, "controller must pin to max");
-        let state = ctrl.state.lock().unwrap();
-        assert!(
-            state.integral.is_finite(),
-            "integrator must be finite under saturation, got {}",
-            state.integral
-        );
-    }
-}
